requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.1
attrs==23.1.0
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==2.0.12
colorama==0.4.6
flake8==3.9.2
flake8-docstrings==1.6.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.6.1
packaging==23.2
pluggy==1.3.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
pytest==6.2.5
pytest-timeout==2.1.0
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tornado==6.3.3
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.18
